rmbase: fold tests into test_pop_rmbase, drop benchmark
